# api/openapi.yaml
openapi: 3.0.0
info:
  title: Vector Search Service API
  description: |
    A vector search service providing semantic search capabilities using PostgreSQL + pgvector.
    
    This service provides document ingestion, embedding generation, and similarity search operations
    designed for integration with Elyra pipelines and other AI workflows.
    
    ## Features
    - Document ingestion with automatic chunking
    - Semantic similarity search
    - Batch processing for pipeline integration
    - Collection management for document organization
    - Asynchronous job tracking
    
    ## Authentication
    This service uses JWT tokens for authentication. Include the token in the Authorization header:
    `Authorization: Bearer <token>`
  version: 1.0.0
  contact:
    name: Vector Search Service
    url: https://github.com/vector-search-service
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://vector-search-service.openshift.apps.example.com
    description: OpenShift deployment
paths:
...
  /api/v1/documents/{document_id}:
    get:
      summary: Get document information
      description: Retrieve metadata and information about a specific document
      tags:
        - documents
      parameters:
        - name: document_id
          in: path
          required: true
          description: Unique identifier of the document
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentInfo'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete document
      description: Delete a document and all its associated embeddings
      tags:
        - documents
      parameters:
        - name: document_id
          in: path
          required: true
          description: Unique identifier of the document to delete
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  document_id:
                    type: string
                  status:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/jobs/{job_id}/status:
    get:
      summary: Get job status
      description: Get the status of an asynchronous job
      tags:
        - jobs
      parameters:
        - name: job_id
          in: path
          required: true
          description: Unique identifier of the job
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/jobs/{job_id}:
    delete:
      summary: Cancel job
      description: Cancel a running or queued job
      tags:
        - jobs
      parameters:
        - name: job_id
          in: path
          required: true
          description: Unique identifier of the job to cancel
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  job_id:
                    type: string
                  status:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Job cannot be cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    DocumentInfo:
      type: object
      required:
        - id
        - collection_id
        - content_preview
        - metadata
        - chunk_count
        - embedding_count
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the document
        collection_id:
          type: string
          description: Collection ID the document belongs to
        content_preview:
          type: string
          description: Preview of the document content
        metadata:
          type: object
          additionalProperties: true
          description: Document metadata
        chunk_count:
          type: integer
          description: Number of chunks in the document
        embedding_count:
          type: integer
          description: Number of embeddings generated
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
    JobStatus:
      type: object
      required:
        - job_id
        - status
        - progress
        - started_at
      properties:
        job_id:
          type: string
          format: uuid
          description: Unique identifier of the job
        status:
          type: string
          enum: [queued, running, completed, failed, cancelled]
          description: Current status of the job
        progress:
          type: number
          format: float
          description: Job progress (0-1)
          minimum: 0.0
          maximum: 1.0
        started_at:
          type: string
          format: date-time
          description: Job start timestamp
        completed_at:
          type: string
          format: date-time
          description: Job completion timestamp
        error_message:
          type: string
          description: Error message if job failed
        result_url:
          type: string
          description: URL to retrieve job results
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Error type
          example: "ValidationError"
        message:
          type: string
          description: Error message
          example: "Invalid request parameters"
        details:
          type: object
          additionalProperties: true
          description: Additional error details
        request_id:
          type: string
          description: Request ID for debugging
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
security:
  - BearerAuth: []
...